requirements.txt
streamlit
numpy
pillow
requests
python-dotenv
groq
chromadb
tqdm
datasets
torch
torchvision
